Smallest window in a String containing all characters of other String
The task is to find the smallest contiguous substring (window) in a given string that contains all characters of another string. For example, in the string "ADOBECODEBANC" with the target "ABC", the smallest window that includes all characters of "ABC" is "BANC". This involves expanding and contracting a sliding window while checking if it contains all characters from the target string.